requirements.txt
streamlit==1.32.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
anthropic>=0.25.0
pydantic>=2.0
weasyprint>=60.0
httpx>=0.27.0
python-dotenv>=1.0.0
jinja2>=3.1.0
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
